Reduce verbosity of the build output
The yum progress bars are useless for automated builds and only make it
harder to see what's happening. The -d1 flag is a somewhat undocumented
way of disabling the progress bar.

The TCL build is also very verbose, mostly due to the fact that it's an
autotools based project. As we're not really interested in the output,
we can disable it with the -q flag for configure and the -s flags for
make. Also added checks for return values to detect errors in the TCL
build.