MXS-4013 Use `hrchy_id` as node id and add `qualified_name`
Each schema node in the sidebar needs to have a unique id. The `hrchy_id`
value is unique, so it's now used as an id. However, this id can't be
used to build the query  as it's not a valid qualified name. So this
adds `qualified_name` attribute for that purpose.