MXS-4362 Split loading and processing of config file
The MaxScale configuration file is now loaded in the main
thread and processed in the MainWorker. That makes it possible
to adjust, if necessary, the startup value of Config::n_threads.
That will have the effect that the created number of routing
workers will be correct from the outset, even if the value
specified in the static configuration file and in some runtime
created configuration file would be different.