Lock `mxs-query-editor` peerDependencies
SkySQL UI don't use `npm ci`, as a result, minor version of
packages are installed. There are breaking changes of
`monaco-editor` in version 0.34 that breaks all current features
of monaco.

This locks peerDependencies to a stable version.
This also removes the `optionalDependencies` and `dependencies` in
the package as they are redundant. Only `devDependencies` and `peerDependencies`
are needed during the build.