MXS-4365 Use worksheet connection type to send kill flag for the query
Before running the queries, the GUI clones the active connection and uses
that to stop the query. This isn't efficient because cloning the connection may
take time and if the user clicks the stop icon before finishing the cloning
process, the Run/Stop icon enters an infinite loading state.

Instead of cloning the current connection, we can use the worksheet
connection without facing busy connection issue as it's only used
for creating new query tabs or disconnecting from the source.

This removes `BACKGROUND` connection binding type.