MXS-3269 Create/destroy user cache at activate/deactivate
When a worker is deactivated, the worker specific data should
be freed. Consequently, when the worker is activated again, it
need to be re-created. This is most straightforward to handle
by deriving from RoutingWorker::Data and by overriding its
[init|finish]_for(...) member functions. This removes the need
for the routing worker to be explicitly aware of data that need
to be inited/finished.

Further, if the startup of MaxScale is changed so that
- first the workers are created,
- then the configuration is created, and
- finally the workers are started,
the distinction between
- thread creation at startup,
- thread creation at runtime, and
- activation of dormant threads,
can be removed and thread initialization can be handled the
same way in all cases.