MXS-3269 Add explicit state
Before this change, the routing worker state was defined
implicitly by its listening and routing states (both boolean).
Primary reason for this is that testing becomes easier if an
_active_ routing worker can be made non-listening, as it allows
for explicit control of which routing worker will get the next
client connection. This change will be followed by other change
that provides that control via the REST-API.