MXS-3269 Report if worker is listening
An active worker will always be listening unless it via the
debug interface of the REST-API has been told not to listen.
A draining or dormant worker will never be listening.

This is one of those fields that will not be reported unless
"verbose" mode has been turned on (when it has been implemented).