MXS-4376: Add /sql/:id/import endpoint
The endpoint will be what starts the import process. As the import
behaves very much like an async query, the results of it are found in
the `/sql/:id/queries/:query_id` endpoint. This allows the same
client-side code to be used for both async query execution as well as
the import management.

Additional information can be placed into the Connection class to let
the client know what state the import is at. If the import needs to be
stopped early, it must be done on the database using a KILL command to
make sure the query execution stops. This means that there's no need to
add a separate endpoint for stopping or otherwise managing the import.