MXS-4372 Render 50 rows for session tables
MXS-4223 made change to show all by default. There could be a large
number of sessions, render all will increase usage of memory which
may crash the browser.

This relieves the usage of memory by limiting the number of
rows the session tables could render at a time. However,
receiving a large JSON still crash the browser, limiting the
size of the JSON respond should be done on the server side.