MXS-4389: Fix schemarouter error handling
The handleError function of the schemarouter ended up calling
clientReply without checking whether the response should be ignored or
not. This caused multiple responses to be sent for a single command
which triggered debug assertions in the client protocol code.

An additional problem that was found as a result of this was the
history_info callbacks stored in the protocol. If an error on the
backend connection causes it to be closed while at the same time a
client sends a new query, it is possible that the callback ends up being
called for a closed connection. This happens because the cleanup was
done in finish_connection() which isn't called until the DCB has been
removed from the zombie queue. The use of a multi-service configuration
and the fact that the schemarouter ended up triggering a fake read event
with the clientReply call were what caused this.