MXS-4376: Allow all JSON types to be created
This allows an empty value of the correct type to be used as a return
value in functions which either retrieve the value or return a default
constructed one.

Added methods for modifying values that can be modified. Since the
Jansson library implements booleans and nulls as unique constant
objects, they cannot be changed from the leaf value and must be replaced
on the parent object. This means that only strings, integers and
floating point values can be updated.