Fix race condition in KILL handling
It was possible for both the KILL response and a hangup to occur on the
same epoll loop which caused events to be queued for both. The code
assumed that the deletion of the LocalClient would prevent this from
happening but in reality both of the events could end up getting queued
to be executed.