Don't call KILL callbacks if session is closed
If the session is already stopping, the router and filter session will
be freed by the time the callback gets called. The client protocol code
that sends the OK packet already checks that the DCB it's writing to is
still open which amounts to the same as checking the session state which
means that it was not affected by this problem. The smartrouter was the
only other component that uses the code in a way where the callback
could cause harm. As both cases survive the case where the callback
isn't called, the solution to this is to simply ignore it if the session
is stopping.