MXS-4376: Add more string helpers
Added mxb::split which splits a string_view value into two string views
at the first occurrence of the delimiter. This makes it easier to parse
key=value arguments.

Added a test case for the string splitting.

Changed the arguments for the uppercase and lowercase functions into
string views. This allows them to also be used with std::string_view.