MXS-4376: Improve ODBC-to-JSON data type conversions
Signed and unsigned integers are correctly converted into their
respective forms. 64-bit integers also now use the SQL[U]BIGINT type
which, although documented to use a long, is really a int on most
systems.

Added code for numeric and decimal types that convert them into JSON
doubles if they were successfully parsed into a double. This might need
to be made optional if a loss of precision occurs due to this
conversion. If the conversion fails, the data type is returned as a JSON
string.