MXS-4376: Return errors for partial results
If a resultset or a multi-statement ends in an error, the error message
was not detected as it was still reported as a success. Moving the error
generation into the Output class takes care of this.

Added the missing warning count to the ok_result() function
signature. This way the behavior should be identical to the one in
MariaDB.