MXS-4376: Add etl JSON configuration
The source and destination servers as well as the tables to be imported
are now read from the JSON payload sent by the client. The source server
connection string is implied by the connection being used and can be
used as-is. The destination server connection string is constructed from
the server entry in MaxScale. Since we know it's guaranteed to be a
MariaDB server, we only need the user to define the credentials that are
used to connect to it.

Removed the UNKNOWN source type as it isn't needed if the validity of
the types is checked before it is assigned. This also makes sure a
partially valid configuration is not possible.