MXS-4376: Add TextResult
Needed for extracting the results of SQL statements that contain the SQL
statements needed to do the ETL operation.

The error state is now reset for every new query which allows the error
methods to be used to detect if a partially successful query was done
without having to store the result.

Added missing getters for the table and schema values in the Table
class. These will be needed for SQL statement generation.