MXS-4404: Make GTID state writing more efficient
Since we know the file needs to be updated often, we can keep the file
handle open. The truncation of the file can also be avoided by writing
the null terminating character into the file. Any trailing data will be
ignored automatically when the data is read into a std::string from a
char buffer.

This change is compatible with the older GTID state file format. It will
still be read correctly even if no null terminating character is found.