MXS-4388: Remove session level LOAD DATA tracking
The LOAD DATA state of the session does not apply to all connections of
the session. An example of this is the fact that slow server might still
be executing session commands when the client starts a LOAD DATA LOCAL
INFILE.

In the client protocol this can be fixed by moving into the LOAD_DATA
state when the mxs::Reply is in the LOAD_DATA state. In this state the
client protocol must also route all incoming traffic to the downstream
component without processing it until the server responds with an OK
packet.

For the QueryClassifier, a call to the new update_from_reply() function
must be done in the clientReply function of any mxs::Component that uses
it. Although incorrect, relying on the global state did not cause any
problems in the QueryClassifier as nothing (hopefully) opens new
connections during a LOAD DATA LOCAL INFILE from a service to another
service. The need to call a second function for the reply to the command
is not ideal. It's sort of easy to forget but luckily only a handful of
modules use it.

For the cache filter, the correct thing to do would be to use the
QueryClassifier class. Since it doesn't use it currently, a simpler
approach is to store the LOAD DATA state in a boolean variable.

Eventually, the result state tracking should not be done by individual
modules but by the mxs::Endpoint that encompasses the router and filter
sessions. This would be an improvement as it would make the result state
tracking work in all routers and filters.

The removal of session_is_load_active() also fixes the problem observed
in MXS-4380 where the backend protocol inspected the session level LOAD
DATA state whenever it was routing a query. This assertion would be hit
when a slow connection executed the session command history during the
LOAD DATA LOCAL INFILE.