Log DCB identity in debug messages
This makes it easier to figure out which DCB is which backend connection
when multiple writes happen in parallel. Using the object address alone
is quite confusing as well and the DCB UUID might be a better candidate.

The debug output is now more useful as the system test framework will
dump the session trace log if the test ends in a hang. With more
detailed debug messages, it should be easier to analyze what really
happened.