MXS-4376: Allow use of /sql connections for ETL
This will make it easier to deal with credentials in the GUI as cookies
can be used to pass the ownership information to the ETL operations. The
only thing that needs to be given is the ID of the connection that is
used as the destination of the ETL operation.

The option to define a target server is still left in as an alternative
approach to starting the ETL without the need to have two separate
connections. This makes it much easier to deal with one-time
fire-and-forget ETL operations that are known to work.

With this change, the separate configuration and creation of the ETL
class is no longer that sensible and it can be combined into one
operation.