Assert with error message in mxb::string_printf
The system errno contains an explanation of what went wrong. This is
useful in situations where the format string itself is valid but the
contents are not. For example, both %s and %S are valid but the latter
accepts only wide character strings and this is only detected when the
formatting is actually done.