MXS-3269 Add state variable for router termination
When going down, we have to know whether a termination of a
routing worker is in process and wait until it has finished
before allowing the shutdown to process.

The termination (to be revealed later) requires message passing
between the RoutingWorker going down and MainWorker, so MainWorker
must remain running until the termination has finished.