MXS-3269 Terminate the thread when worker deactivated
When a worker is deactivated, if its index is the largest the
thread will also exit.

The termination is handled so that
- first the router is removed from the living (this_unit.ppWorkers),
- then it stops receiving messages,
- then waits for TERMINATION_DELAY (currently 5 seconds),
- and finally arranges for it to be deleted by the MainWorker.

The critical point is broadcasting, because its done not only from
from the MainWorker, but from other locations as well. Broadcasting
essentially is:

    1) auto nWorkers = this_unit.nRunning.load(...);
    2) for (int i = 0; i < nWorkers; ++i)
    3) {
    4)    this_unit.ppWorkers[i]->execute(...);
    5) }

This will fail if:
- T1 calls 1 and gets nWorkers and is then put to sleep.
- The number of workers is reduced.
- It takes longer than TERMINATION_DELAY for T1 to reach
  the end of the this_unit.ppWorkers array so it will at
  that point access a pointer that no longer points to data
  or is NULL.

Considering that Worker::execute(...) only places a task on the
worker message queue, it's quite unlikely that could ever happen.

If multiple workers are deactivated, they are terminated one at a
time from the end. Makes things simpler and ensures that
RoutingWorker::Data::finish_for(RoutingWorker*) will be called
in the inverse order. That requires the call to be moved from the
deactivating phase to the termination phase, which is not yet done.

Further, there no longer is a need for this_unit.nCreated, so it
will be removed in a forthcoming change.