MXS-4407: Add Extractor class
The Extractor class implements the API needed to coordinate the ETL
operation. The two main responsibilities of it are:

* To synchronize the dump threads with the coordinator thread so that
  the tables being imported are consistent.

* To generate SQL statements that are used to create the tables and to
  load and insert the data in the destination MariaDB server.

The Source enum is no longer needed as the concrete class can be created
when the configuration is being parsed. This means that the
Configuration class now holds only the source and destination ODBC
connection string but other options (e.g. for automatically dropping
existing tables) still need to be added there.