MXS-3269 Clear rworker data only in RoutingWorker::post_run()
Since workers/threads now really terminate at runtime, it is
unnecessary to clear the data when a thread is deactivated and
only do that when the thread is going down.

For the same reason some initialization can be moved from
RoutingWorker::activate() to RoutingWorker::pre_run().

With this change, the finish_for(RoutingWorker*) will also always
be made in the inverse order of how the init_for(RoutingWorker*)
calls were made.