MXS-4270 Alter crypto_sign() to work with any length password
The reference implementation assumed that a specific length
private key is given to crypto_sign() and that the public key
is given with the private key. Use the password as key
instead and generate the public key as in crypto_sign_keypair().
The changes here are similar to the ones used in the client
authenticator code in MariaDB Server.

Also adds a README-file.