MXS-4407: Make ETL concurrency configurable
The `"threads"` field can be used to set the number of tables that are
imported in parallel. By default at most 16 tables are imported in
parallel. `"threads"` values are clamped to the number of tables being
imported.

Also added a check that prevents an ETL to take place which would import
no tables.