MXS-3269 Disallow #thread-change, when change is in process
If a change could be made when a termination is progress, then
there could be the situation that the init_for() function is
called by the new thread, before the finish_for() function is
called by the terminating thread. Since a change in the number
of threads is not something that should be done willy-nilly,
it's not unreasonable to require that one change has finished
before the next one is made.