MXS-4407: Implement the actual ETL operation
The code can now import tables from MariaDB servers to other MariaDB
servers.

Currently, an error on one table does not cause the processing of other
tables to stop. This could be changed so that the first error causes the
rest of the tables to be ignored. This would speed up how fast errors
are reported but it would also mean that subsequent errors in later
tables would not be found.