MXS-4407: Hide Connector-C symbols
If the symbols are not hidden, the MariaDB ODBC connector would end up
using the symbols from the static library that MaxScale uses instead of
the system libraries that it was linked against.

Not exporting the symbols at all caused some problems with the
mysql_change_user behavior in tests. It somehow broke the connector's
builtin plugins which no longer resolved to valid symbols. For this
reason, the symbols are only excluded from the maxscale-common library
which links against unixODBC and left intact elsewhere. This appears to
prevent the MariaDB ODBC driver from seeing the symbols in the
maxscale-common library.

For some unknown reason, the link order of maxscale-common must also be
changed to include the static authentication and protocol modules before
any of the dynamic libraries. Without this, the OpenSSL libraries would
not link correctly for other libraries that do depend on the connector.