MXS-3269 Adapt rworker_basics to 'termination_in_process'
The number of threads can only be adjusted if termination is not
currently in process. Just checking whether it is in process is not
sufficient, however, but when decreasing the number of threads, it
must be ensured that all existing threads are "Active".

The basic problem is that since the curl executable is used, it is
not easy to do error-handling, so it is better to arrange things so
that no curl-operations will fail (but for some real error).