MXS-4405 Rename all occurrences of the keyword `session` or `sess`
`session` is not an accurate name as it indicates the period of time between
a client connecting to and the client disconnecting from the database.
In addition, the UI uses `Query Tab` instead of `Query Session` and the
file name of the newly ORM entity for `querySession` is now `queryTab`.