MXS-4407: Add ODBC-375 workaround
Wrapping the result in a BEGIN NOT ATOMIC block prevents the crash but
it causes the result to be arranged differently: all OK packets are
combined into one trailing OK packet, resultsets with a trailing error
appear as an empty resultset and syntax errors in one part cause the
whole SQL statement to be rejected.

All of this is still better than crashing the whole process.