MXS-4405 Adjust model relationships
QueryConn entity relationship between Worksheet and QueryTab entities
can be inversed, .i.e, one to one inverse relationship. This makes
it easier to bind connections to an entity.

A connection can belong to 0 or 1 Worksheet, or it can belong
to 0 or 1 QueryTab.

With this inverse relationship approach, we can later create new
connection with FK for a different entity, .e.g. ETL, Data Compare...

Validate connection and cleans up data can also be done easier. If
a connection is expired but its persistent data in the GUI is still
there, call `QueryConn.cascadeRefreshOnDelete(id)` will delete
the connection persistent data and refresh either Worksheet or
QueryTab data.