MXS-4421: Deep-copy all client packets
This prevents the query classification data from being shared by packets
read at the same time. In practice this only happened when the client
sent more than one packet at a time which usually means some form of
batch operation was being done.

This also fixes the debug assertion and the info level message in
readwritesplit that would otherwise break. This is actually caused by
MXS-4419 but the immediate problems it causes are fixed by this commit.