Add custom timeouts for tests
The tests can now exceed the hard-coded 300 second default timeout for
tests without having to repeatedly call it. This useful when a single
command takes a long time to complete and there's no opportunity to call
reset_timeout() in the middle of it. In practice the only case where
this is needed is when an external program implements the actual test
case and the test framework is just used to start it.