MXS-4405 Add `editorsMem` state module
To use ORM syntax, all state modules related to ORM entities should
be registered via vuex-orm.

If the `editors` state module is registered via vuex-orm, it will be a key
under `ORM_NAMESPACE`, hence, all states in it will be persisted. However,
there are some states in the `editors` module that shouldn't be
persisted as those are not relevant to ORM.
 e.g. `selected_query_txt, file_dlg_data, charset_collation_map,...`

This creates `editorsMem` state module to keep those states in memory.