MXS-4405 Add a workaround for persisting `FileSystemFileHandle`
It turns out that vuex-persist can't deal with `FileSystemFileHandle`
as state data is serialized with `JSON.stringify`.
This problem doesn't exist with the sync methods for normalizing objects
in 22.08, somehow it is ignored in vuex-persist. But with the ORM models,
objects are not nested deeply, `JSON.stringify` may have work well which
causes the loss of `FileSystemFileHandle` after refreshing the browser.

This creates `fileSysAccess` state module to manually store file handle
to IndexedDB via `localForage`.

This also moves `initEntities` call back to `queryEditorConfig` module
as now there is a need to dispatch `fileSysAccess/initStorage`.