MXS-4424: Fix the order of fields in indexes
The fields weren't ordered in the correct way which caused them to be
added in the order the JOIN happened instead of in the order the index
defines them in.

Removed the unnecessary join with pg_am that was to be used to determine
which type of index to create. Since the special indexes in PostgreSQL
can't be converted into MariaDB ones, the information isn't needed
anymore.