MXS-4424: Add support for foreign keys
The foreign keys are declared using the syntax that defines both the
constraint name as well as the index name. If the index for the foreign
key is the primary key, the index name is ignored by MariaDB. This is
different from PostgreSQL where even the primary key has a name.