MXS-4437: Expose ODBC catalog functions
These will be used to implement a GenericExporter class that can be used
to migrate data from any ODBC source.

The results from a call to a ODBC catalog function is a resultset of a
known width. The drivers are allowed to add columns and the column names
aren't guaranteed to be the same. ODBC API documentation recommends
using the field numbers to extract the information.