MXS-4383 Make initial preparations
The prototype

  vector<string_view> qc_get_table_names(GWBUF* stmt, bool fullnames);

will be changed to

  struct QcTableName
  {
      string_view db;
      string_view table;
  };

  vector<QcTableName> qc_get_table_names(GWBUF* stmt);

to remove the need for finding the qualified name in the statement
as it will not be found if it was expressed like "`db`.`table`".
Otherwise it is also much better to express all information at once
instead of having to make two calls or look for "." in the name.

With this change, qc_sqlite and qc_mysqlembedded will internally
store the data in a manner suitable for the prototype change, but
externally the API is still the same. Due to that, the fullnames
must still be generated and stored separately.