MXS-4437: Add GenericExporter
The GenericExporter uses the ODBC catalog functions to create the
`CREATE TABLE`, `SELECT` and `INSERT` SQL statements.

The `CREATE` statement will have comments in it that will display the
native type name that was reported by ODBC. This should help the
end-user determine whether the type that was deduced from the ODBC SQL
type is close enough or whether it needs to be modified.

This could be used as the starting point for all Exporter classes but
the benefit of doing this isn't that great: a specialized `CREATE`
implementation most likely also needs specialized `SELECT` and `INSERT`
implementations.

Since the GenericExporter has to use constants from the ODBC API, the
implementation is in a separate file. This isolates any problems that
sql.h and sqlext.h might cause into a single file.