MXS-3269 Add RWorker::Data post start initialization
Classes that have worker specific data must be derived from
RoutingWorker::Data, which causes their init_for() function
to be called when the worker is started.

That function must be called also when such an object is
created when the routing workers have already been started.

That cannot be handled by RoutingWorker::Data, but the derived
class must call Data::initialize_workers() in its constructor.