MXS-3269 Run server/core unit tests by MainWorker
Otherwise some assertions will fire.

The shutdown code of MainWorker had to be tuned as well, as
currently some activity is started in gateway.cc, yet stopped
in MainWorker. That needs to be made symmetric; either something
is started and stopped by MainWorker or started and stopped by
gateway.cc.