MXS-3269 Replace qlafilter with hintfilter in test_filter
The test only tests runtime creation of filters and for some
reason, creating a qlafilter is very slow. As the point of the
test is not to test qlafilter, it is replaced with hintfilter
that makes the test run in a fraction of the time it otherwise
takes.

The reason for the slowness of qlafilter should be figured out,
but that is a separate problem.