MXS-4440: Fix idleness calculation of connections
The is_idle() function needs to be backported from 6 into 2.5 as the
upper level code needs to know whether all connections are idle or
not. Only the backend side of it is needed and the client-side can be
left unimplemented: it is only used when idle_session_pool_time is
enabled and this was added in version 6.

In 2.5 implementing the idleness tracking using only the backend
protocol side is a simpler approach as it uses the resultset state
tracking which is known to work and exists in 2.5. In version 6 this
could be done on the client side as it tracks the number of expected
responses but the approach used here is functionally the same. The only
semantic difference is that if something causes a backend to not be idle
that wasn't triggered by the client, it could end up generating a ping
that would otherwise not be sent. Once this commit is merged upwards,
the code can be changed to use the client protocol's is_idle() method.