MXS-4440: Fix idleness calculations in 6.4
This can be fixed by just extending the check to also make sure no
results are expected.

The lack of this check might also indicate that the runtime filter
alterations for sessions might also suffer from problems. If true, the
filters would get a clientReply call without a matching routeQuery call.

The test case from 2.5 verifies that this fix solves the problem.