MXS-4437: Make ETL operations cancellable
The same mechanism that is used to cancel a query can be used to cancel
the ETL operation. Even if the query on the backend server is not
canceled by the SQLCancel call, the processing of any results stops
immediately after the latest batch has been inserted. At this point the
transaction rolls back and an error is reported.