MXS-4442: Fix handling of GnuTLS certificate chains
The documentation of the gnutls_pcert_list_import_x509_raw function
states that a list of objects should be used. What exactly a list is was
a bit ambiguous. It turns out that it required an array of
gnutls_pcert_st structs instead of a single gnutls_pcert_st struct.