MXS-4443 Implement ETL task model
ETL tasks will be persisted to IndexedDB.
Each task is an object with the following attributes:
- id: UUID
- name: The users provide the name for the ETL task or it will be generated
automatically.
- status: Enum values (running, canceled, error, done)
- sql_script: Generated script (the response of `/sql/:id/etl/prepare`)
- active_stage_index: Index of the stage that the ETL task is in.

This model is a 1 to many relationship with `QueryConn` model. In
other words, an ETL task would have 2 connections, a source connection
via ODBC and a destination connection to a MariaDB server in MaxScale.