MXS-4366: Track execution time of async queries
The query execution time is now always included in the response from the
GET /sql/:id/queries/:query_id endpoint. This allows the client
application to track how long it has been running.

This also fixes a data race where the SQL being executed was written to
and read from without any locking. In practice this wasn't a likely
problem but it could happen in theory.

Moving the concrete query implementation into a protected function that
the derived classes implement keeps the API stable and also makes it
easier to do common tasks there without having to implement it in all of
the deriving classes.