MXS-4445 Remove `type` field in `QueryConn` model
It's no longer needed since the removal of using URL to determine
active worksheet.

This also clearly specifies which attributes of the worksheet connection
to be kept after cloning it for query tab instead of shallow merging
the entire worksheet connection object. The connection's `name` and meta
fields will be kept.