MXS-4439: Never abandon events
This prevents lost messages when the target message queue is full. This
also avoids the problems that would be caused by a message being lost
(memory leaks, hanging clients etc.).

In 22.08 this problem no longer exists as the messages are passed using
a different mechanism. There is, however, a problem with an increase in
memory usage whenever a peak in activity is seen.

A warning is still logged if the worker is unable to receive a message
in under five seconds. The warning now also includes a suggestion to
disable reverse name resolution.