MXS-4440: Fix pinging of idle connections
If the client has been idle for less than the configured limit but some
of the backends have been idle for longer, then the backends still need
to be pinged. Only if the limit is exceeded must the idleness of the
backends be inspected.

This fixes the connection_keepalive test.