MXS-4454: Use known good targets for secmd repliers
This avoids the problem where a server with no known databases on it is
selected as the backend from who the reply is returned to the client. If
the backend is empty because the user doesn't exist there, the session
command would end up failing.

Iterating over the shard map is not ideal as the worst-case complexity
is O(nm) where n is the number of tables and m is the number of
targets. Usually this should result in a match on the first comparison
of a mxs::Target and a SRBackend.