MXS-4445 Persist ETL connection name to EtlTask model
After finishing an ETL task, source/destination connections will be
closed as there is no need to keep it open. Howerver, it would be
informative for the users to have information about the connections
that were opened for finished ETL tasks.

This adds `meta` field to EtlTask model for storing connection name.