Fix dependency packages for installing unexpected incompatible version of vue
After regenerating the `package.json`, the `vuetify-loader` package was
installed with version `1.9.2` where it specifies the minimum version of vue
is 2.7 as its peerDependencies. This causes problem with running unit tests
via `@vue/test-utils` package.

This replaces the caret `^` with `~` for vue, vue-template-compiler, vuetify
and vuetify-loader packages so it won't install any unexpected versions.