MXS-4440: Fix alteration of filters for sessions
The alteration of filters used the is_idle() function which now never
returns true inside the routeQuery function. Previously the function
technically did the wrong thing as the protocol would no longer be idle
once it had passed the query to the session's routeQuery. Checking
whether older responses are expected is a better approach.

Added a test case that revealed the problem.