Reorder checks in mysqlmon_failover_stress_ss
The time it takes for replication to start can apparently be longer than
one second, especially as the servers are started up moments before the
check for the number of clients is done. Reordering the checks allows
the client count check to take place last which should make it more
likely to succeed.