MXS-4430 Restructure queryEditor dir
The query editor will have new views in the future like ETL, data comparison
and ERD. Even those features are not really relevant to `Query Editor`, they
will be a part of `mxs-query-editor` package. So, this restructures the location
of some components to make it easier to extend to have new views.