MXS-4107 Accept 0 nested params for non-nested param
Not providing nested parameters to a parameter not accepting
nested parameters is ok. Whether or not the caller is interested
in unreocognized parameters is not what decides the validity of
the operation.

With the above change, the decision as to whether the validation
of nested parameters failed can be made logical; it failed if the
validation of the nested parameters failed or if there were
unrecognized parameters.