MXS-4460: Check connection state in delayed routing
The Routable itself cannot be used to determine if the Endpoint is still
open as closing a ServiceEndpoint will free the router session. For this
reason the Endpoint itself must be stored in the Routable. The correct
thing to do would be to pass the Endpoint reference during construction
so that the router's could pass that down to the Routable but that would
require a lot of changes in multiple places. This should be refactored
to do this in the next major release.

This also partially fixes the problem where a router session did not
know which service it belonged to when it was a part of a multi-level
service-to-service configuration. It still doesn't know it during the
object construction but it will be available during routing via the
Routable::endpoint() function.