MXS-4454: Use known good targets for TARGET_ANY
This avoids the same problem that the session commands had which would
end up causing queries to be routed to backens that have not been
successfully mapped.

The complexity cost of "inverting" the hashtable is a lot compared to
the small memory overhead of maintaining a std::set of pointers. This
allows the worst-case complexity to be reduced to O(n log n). In
practice this will be a trivial cost as the number of backends will
always be much smaller than the number of tables in them.