MXS-4446 Extend `getChildNodeData` to control `activatable` node
By default, VIEW and TBL nodes are activatable. i.e. the node is
highlighted with pale blue background to indicate that the node is
being "previewed" with top 1000 rows. With ETL schema tree, this
behavior is not needed, so it should be disabled.

This adds `activatable` parameter to conditionally override the default value.

This also renames `canBeHighlighted` to `activatable`.