MXS-4467: Prefer previous target for TARGET_ANY
This makes the behavior of TARGET_ANY queries more predictable.

This also fixes the problem where a session with autocommit disabled
won't commit a single-shard transaction correctly when a COMMIT command
is received. This is simply a temporary side-effect and a follow-up
commit will be added that routes all transaction related commands to all
backends.