MXS-4467: Route trx control commands to all nodes
This will cause the transaction state to be identical on all nodes that
the schemarouter uses. The downside of this approach is that it also
involves any nodes that do not participate in the transaction which
causes extra traffic that wouldn't technically be needed if the
transaction targets only one shard. Routing hints can still be used to
direct the commands to the correct location.