MXS-4471: Select only valid targets
The target selection for single-target queries did not check the state
of the backend connection before choosing the target server. This caused
the query to fail if the first server in the std::set was not one the
session was connected to.

In addition to the possibly of choosing a backend that the session is
not connected to, the order in which nodes were chosen was not
deterministic. By iterating over the backend server and checking if they
are found in the std::set, the order is now deterministic with respect
to the order in which the server are declared in the configuration.