MXS-4470: Route database changes to relevant backends
This way the behavior for both the initial default database and any
later changes to it work the same way.

However, this does not preven the problem where, upon re-execution of
the session command history, the connections to backends that do not
have the database are discarded as they are unable to successfully
execute the command.