MXS-4474: Use lcalls to deliver fake events
Skipping the message queue for fake events that end up on the same
worker prevents the message queue from filling up. This also reduces the
pressure on the whole message queue as fake read events are a common
event in MaxScale 6.4: they are used to wake up the client protocol
while session commands were being executed.