MXS-4474: Prevent hangs due to KILL commands
The execution of a KILL command involves the MainWorker which can end up
blocking. This blocking happens when the MainWorker is waiting for the
RoutingWorkers while they are collecting the connection information. If
enough clients execute a KILL command at the same time, some of the
Worker::execute() call will inevitably block.

Due to the recent changes to the message queue where events are never
abandoned, the KILL execution can end up blocking the whole MaxScale
process. To prevent this, events must be abandoned in certain
cases. Luckily the cases where these appear to happen are either rare or
not critical and the cases where they are, the code is easily modified
to handle it.